backend/main.py
from fastapi import FastAPI
from pydantic import BaseModel
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

# Placeholder for database connections
# QDRANT_URL = os.getenv("QDRANT_URL")
# NEON_DB_URL = os.getenv("NEON_DB_URL")

class Query(BaseModel):
    text: str

@app.get("/")
def read_root():
    return {"message": "Backend is running"}

@app.post("/query")
def query_rag(query: Query):
    # Placeholder for RAG logic
    # 1. Embed the query.
    # 2. Search Qdrant for similar documents.
    # 3. Get content from Neon DB based on search results.
    # 4. Generate a response using an LLM.
    return {"text": query.text, "answer": "This is a placeholder answer."}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
