backend/requirements.txt
fastapi
uvicorn
qdrant-client
psycopg2-binary
python-dotenv
